requirements_client.txt
tqdm
feedparser
faker
pyautogen==0.2.31
numpy
requests
tqdm
bs4
networkx
pillow
flaml[automl]